🧹 Code Health: Extract sweep logic from Nec2Engine
Extracted the adaptive SWR sweep logic into its own module `src/physics/sweep.ts`. This reduces the size and complexity of the core `Nec2Engine` class, making the physics logic more maintainable. Constants were moved to `src/physics/constants.ts` to keep the code DRY. Tested thoroughly with the existing suite.
